+++ lib/src/_layout.dart
@@ -5,12 +5,12 @@ import '_base.dart';
 // https://bulma.io/documentation/layout/container/
 
 class Container extends BulmaComponent {
-  final List _children;
-  Container({List children}) : _children = children;
+  final _content;
+  Container({dynamic content}) : _content = content;
 
   @override
   Element createElement(BuildContext context) {
-    return new Element('div', classes: ['container'], children: _children);
+    return new Element('div', classes: ['container'], content: _content);
   }
 }