+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 0.0.2-dev
+## 0.1.0
 
 - `Input.value` to provide the current value in the `<input>` field.
+- Upgraded to `domino` `0.2.0`, with **breaking changes**: `content` is used instead of `text` or `children`.
 
 ## 0.0.1